Add tests for custom link feature
Added some new tests in the 'graph.helper.test' suite to check whether
the customization of the color for a certain link is working correctly.

We will be checking whether the stroke color matches the expected one in
both scenarios: when a custom color is set and when no color has been
passed along the link data.