fix: Graph fails to render without config prop (#84)
When the user didn't add the 'config' prop to the 'Graph' component (which should be possible, as it's an optional field), the graph failed to render whenever its data changed.

The component would work the first time around, but when it received new data, it threw an error as described [in this issue](https://github.com/danielcaldas/react-d3-graph/issues/81) and the view would not update.

The root of the problem was located in the `componentWillReceiveProps` method of the `Graph` component. When defining the constants `configUpdated` and `transform`, the case in which `newProps.config` was set to either null or undefined was not being considered. This caused errors when calling `isDeepEqual` and also when trying to get the `panAndZoom` property during `transform` initialization.