feat: Customize nodes with JSX views
A new 'view' property can now be passed down to graph nodes. This new
property will receive a view, which can include plain HTML or even
React components.

Please, note that this is only available as a low level property in
individual nodes and
cannot be defined in the config object. This feature is mostly thought
for dynamic content, so 'svg' property should be enough to cover any
scenario in which the nodes' content don't change depending on their
inner data.

Also, if the node has values for both 'svg' and 'view' properties,
'view' will always prevail. If we want 'svg' to be applied instead, we
must not assign a value to 'view'.