feat: Focus graph on a node
It is now possible to set a "focusedNodeId" to the graph, which will
result in the selected node appearing on the center of the displayed
area of the graph.

Optionally, aside from centering the view on it, a certain zoom can be
applied while the node is focused on screen. No zoom is applied by
default, but that behaviour can be configured by using the new
"focusZoom" parameter in the config object. This setting should receive
a number with the zoom level to apply, so the larger it is, the bigger
the node will appear on screen.