fix: Wrapper for node's custom view
Inside of the 'foreignObject' tag, there is now a body tag which will
wrap around the defined custom view. This wrapper has the same dimensions that were set as node size in the
config object.

Because this wrapper has explicit dimensions now, the developer will be
able to use relative sizes (such as percentages) in their custom view's
parent element.

This fixes an issue where the custom component's height
and width were set to 100%, and it failed to retrieve the size for the
'foreignObject' tag. Having a real parent in the embedded HTML solves the problem, as our new view is no longer unable to retrieve its dimensions.