feat: Custom view sample added to sandbox
The sandbox includes now a new example where the custom view property is
demonstrated.

In this case, each node represents a person and, for every one of them,
an additional set of data is included. The defined custom view displays
all of this data: name, gender and whether they're able to drive a bike
and / or a car.

The values for the 'customView' property are generated dynamically in the
'custom-node.data.js' file.

In order for the user to watch this example, they must enter the sandbox
with the following query param:

    {default sandbox url}?data=custom-node