fix: Error unmounting nodes with custom view in sandbox example
When trying to remove nodes in the sandbox example for the custom view
feature, the application threw the following error.

```
<body> tried to unmount. Because of cross-browser quirks it is
impossible to unmount some top-level components (eg <html>, <head>, and
<body>) reliably and efficiently. To fix this, have a single top-level
component that never unmounts render these elements.
```

This was introduced when I added a body tag as a wrapper for the
embedded code that serves as a custom view. Replacing that tag with a
lower-level one fixed the issue.