feat: View generator function to customize nodes
The `view` property where we could assign an already generated view to
our nodes has been replaced by a new property `viewGenerator`.

This new property can be defined both at the global configuration object or at node
level, and it will consists on a function that receives a node object and uses its data to return a
view:

```
viewGenerator: (node) => <CustomViewComponent node={node}
```

With the previous method, the node view was not automatically updated
whenever the node data changed. The only way for this to happen was by
manually overriding the value of the the `view` property.

With the current approach, there are a couple of advantages:

* The view generator can be defined in the config object, so we don't
have to set a property for the view in every node. All of the nodes will
share this global setting.

* The library will be in charge of rendering the node custom view
whenever the data changes, as the generator method receives the current
node data as a parameter and returns a view generated using this values. This means the developer doesn't have to worry about updating a view property on each node everytime their data changes, as the view will be automatically updated anytime the Node component `render` method is called.

The sandbox example for this feature has also been updated to reflect the new
changes.