refactor: Inner property to override view generator
When building a node component props, a new property called
`overrideViewGenerator` is added. This property is set to `true` when
there's a svg provided for an specific node which doesn't have its own
view generator. In this case, we assume the svg should have precedence
over the default view generator set in the global settings.

In any other scenario, the view generator will always be applied instead
of the svg. Only when `overrideViewGenerator` is `true` will we assume
the `svg` property to be more important.

Before, this was done by setting the `viewGenerator` property to `null`,
but the new property makes the code more readable.