test refactor: Implicit values inherited from node
Since some of the recently added values are directly inherited from the
`that.node` data object and `buildNodeProps` does not alter them in any
way, they are no longer hardcoded in the test `toEqual` check.

The object `that.node` contents are added via the spread operator
instead, which will make things easier if the sample data changes at some point.