fix: Add `viewGenerator` to default config
When installing the library in a project, the `viewGenerator` function
defined in the config object by the user did have no effect. The reason
for this was that properties not included in the default config object
(graph.config.js) were been ignored when building the graph and
rendering it in the browser.

In order to solve this issue, a `viewGenerator` property is now included
in the default config object, with an initial value of `null`.