fix: Don't show function or null properties in sandbox
The current sandbox example created a form that allowed all of the
available config properties to be customized in real time.

The `viewGenerator` property, though, cannot be customized with a form.
Since its value must be a function, the user cannot set it via a form
field.

Also, the fact that it's initialized to null made some of the sandbox
examples to throw an error because the form schema generator doesn't
support neither null or undefined values. Functions are not allowed
either, and in this case an error would appear on screen by saying that
this type of data is not supported.

The current change filters the config elements so that neither null,
undefined or function values are included in the config form.