fix: Consider hidden links when obtaining matrix
The link matrix was not been properly updated when the collapsible
setting was enabled. This is because when obtaining its values it didn't
check whether the different links were hidden or not.

This change fixes the issue and prevents the link matrix from containing
wrong values which cause orphan nodes are not hidden properly during
collapse.