fix: Reset transition duration after focus animation
The transition duration property was been applied to the graph container
at all times, which affected to every other transformations that
triggered on it. This included the default zoom or the drag and drop
functionality, among others.

In order to fix this, that transition duration value is set back to zero
after the focus and zoom animation has finished. Some actions on the
node do also cancel the animation in case it's still ongoing, like
trying to drag a node.

With this change, there's no longer a delay when moving a node to
another location or while performing a zoom.