fix: Drag & drop in focused node
There was an issue with the drag and drop feature when used along the
focus and zoom transformation.

The problem was that we were recalculating the transformation everytime
the page was rendered, which meant that a new transform value would be
created anytime the node moved to a new position.

This issue only affected the focused node itself, since the
`getCenterAndZoomTransformation` did only depend on that node
coordinates.

To avoid abnormal behaviours, the transformation is only obtained when the graph receives props. By doing this, the current view position is preserved until a new focused node is set. If the user drags the node around, current zoom and position will be unaffected by it. This matches the default behaviour we encounter in any other scenario, when moving any other nodes or not having activated the focus and zoom feature.