refactor: Reset transition duration on drag start
The `enableFocusAnimation` property in Graph' state is now set to
`false` in the `dragStart` event instead of during `dragMove`. This
change should have a positive impact in the application's perfomance, as
the `enableFocusAnimation` value will only checked once instead of being
read in every position change during a drag operation.