fix: Node not centered when focusing with zoom
One of the latest changes broke the focus feature in certain scenarios.
When a focus zoom different from one was applied, the view was not
properly centered in the selected node. The transformation obtained in
the 'getCenterAndZoomTransformation' did not work as expected. Applying
a single translate before scaling didn't seem to do the trick, and the
second translate is necessary for the view to display the selected node
in the center after scaling the graph.