fixing bug:
Once the user passed FocusNodeId to the graph, and he moved the graph to a new position, if componentWillReciveProps will call again the graph will always return to the location of the Focus Node.
On the other hand, if the user cancels the focus node and passes an undefined parameter to the graph, the graph will return to the initial position and will be a jump in the position when we will render again the Graph component.
We want the graph to always keep its last position - when there is a focus node and when there is no, and when we will render the component again (componentWillReciveProps), the graph will not change its position unless we passed focus node id.